requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
twilio==8.3.0
openai==0.28.0
Pillow==10.0.0
requests==2.31.0
gunicorn==20.1.0